Move PreBuiltPage specific helper methods to its helper file
